Extract large function to new file
